refactor: consolidate pool implementation (#1868)
