+++ x/gamm/pool-models/balancer/marshal_test.go
@@ -1,4 +1,4 @@
-package balancer
+package balancer_test
 
 import (
 	"encoding/hex"
@@ -8,10 +8,12 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/require"
 
+	"github.com/osmosis-labs/osmosis/v7/x/gamm/pool-models/balancer"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-var ymlAssetTest = []PoolAsset{
+var ymlAssetTest = []balancer.PoolAsset{
 	{
 		Weight: sdk.NewInt(200),
 		Token:  sdk.NewCoin("test2", sdk.NewInt(50000)),
@@ -25,7 +27,7 @@ var ymlAssetTest = []PoolAsset{
 func TestPoolJson(t *testing.T) {
 	var poolId uint64 = 10
 
-	jsonAssetTest := []PoolAsset{
+	jsonAssetTest := []balancer.PoolAsset{
 		{
 			Weight: sdk.NewInt(200),
 			Token:  sdk.NewCoin("test2", sdk.NewInt(50000)),
@@ -35,7 +37,7 @@ func TestPoolJson(t *testing.T) {
 			Token:  sdk.NewCoin("test1", sdk.NewInt(10000)),
 		},
 	}
-	pacc, err := NewBalancerPool(poolId, PoolParams{
+	pacc, err := balancer.NewBalancerPool(poolId, balancer.PoolParams{
 		SwapFee: defaultSwapFee,
 		ExitFee: defaultExitFee,
 	}, jsonAssetTest, defaultFutureGovernor, defaultCurBlockTime)
@@ -48,7 +50,7 @@ func TestPoolJson(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, string(bz1), string(bz))
 
-	var a Pool
+	var a balancer.Pool
 	require.NoError(t, json.Unmarshal(bz, &a))
 	require.Equal(t, pacc.String(), a.String())
 }
@@ -58,7 +60,7 @@ func TestPoolProtoMarshal(t *testing.T) {
 	decodedByteArray, err := hex.DecodeString("0a3f6f736d6f316b727033387a7a63337a7a356173396e64716b79736b686b7a76367839653330636b63713567346c637375357770776371793073613364656132100a1a260a113235303030303030303030303030303030121132353030303030303030303030303030302a110a0c67616d6d2f706f6f6c2f3130120130321e0a0e0a05746573743112053130303030120c313037333734313832343030321e0a0e0a05746573743212053530303030120c3231343734383336343830303a0c333232313232353437323030")
 	require.NoError(t, err)
 
-	pool2 := Pool{}
+	pool2 := balancer.Pool{}
 	err = proto.Unmarshal(decodedByteArray, &pool2)
 	require.NoError(t, err)
 
@@ -67,7 +69,7 @@ func TestPoolProtoMarshal(t *testing.T) {
 	require.Equal(t, pool2.PoolParams.ExitFee, defaultExitFee)
 	require.Equal(t, pool2.FuturePoolGovernor, "")
 	require.Equal(t, pool2.TotalShares, sdk.Coin{Denom: "gamm/pool/10", Amount: sdk.ZeroInt()})
-	require.Equal(t, pool2.PoolAssets, []PoolAsset{
+	require.Equal(t, pool2.PoolAssets, []balancer.PoolAsset{
 		{
 			Token: sdk.Coin{
 				Denom:  "test1",
