+++ x/gamm/pool-models/balancer/pool_params.go
@@ -1,0 +1,78 @@
+package balancer
+
+import (
+	"errors"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/osmosis-labs/osmosis/v7/x/gamm/types"
+)
+
+func NewPoolParams(swapFee, exitFee sdk.Dec, params *SmoothWeightChangeParams) PoolParams {
+	return PoolParams{
+		SwapFee:                  swapFee,
+		ExitFee:                  exitFee,
+		SmoothWeightChangeParams: params,
+	}
+}
+
+func (params PoolParams) Validate(poolWeights []PoolAsset) error {
+	if params.ExitFee.IsNegative() {
+		return types.ErrNegativeExitFee
+	}
+
+	if params.ExitFee.GTE(sdk.OneDec()) {
+		return types.ErrTooMuchExitFee
+	}
+
+	if params.SwapFee.IsNegative() {
+		return types.ErrNegativeSwapFee
+	}
+
+	if params.SwapFee.GTE(sdk.OneDec()) {
+		return types.ErrTooMuchSwapFee
+	}
+
+	if params.SmoothWeightChangeParams != nil {
+		targetWeights := params.SmoothWeightChangeParams.TargetPoolWeights
+		// Ensure it has the right number of weights
+		if len(targetWeights) != len(poolWeights) {
+			return types.ErrPoolParamsInvalidNumDenoms
+		}
+		// Validate all user specified weights
+		for _, v := range targetWeights {
+			err := ValidateUserSpecifiedWeight(v.Weight)
+			if err != nil {
+				return err
+			}
+		}
+		// Ensure that all the target weight denoms are same as pool asset weights
+		sortedTargetPoolWeights := SortPoolAssetsOutOfPlaceByDenom(targetWeights)
+		sortedPoolWeights := SortPoolAssetsOutOfPlaceByDenom(poolWeights)
+		for i, v := range sortedPoolWeights {
+			if sortedTargetPoolWeights[i].Token.Denom != v.Token.Denom {
+				return types.ErrPoolParamsInvalidDenom
+			}
+		}
+
+		// No start time validation needed
+
+		// We do not need to validate InitialPoolWeights, as we set that ourselves
+		// in setInitialPoolParams
+
+		// TODO: Is there anything else we can validate for duration?
+		if params.SmoothWeightChangeParams.Duration <= 0 {
+			return errors.New("params.SmoothWeightChangeParams must have a positive duration")
+		}
+	}
+
+	return nil
+}
+
+func (params PoolParams) GetPoolSwapFee() sdk.Dec {
+	return params.SwapFee
+}
+
+func (params PoolParams) GetPoolExitFee() sdk.Dec {
+	return params.ExitFee
+}
